requirements.txt
flask==3.0.0
google-generativeai==0.5.2
numpy==1.26.4
python-dotenv==1.0.1
scikit-learn==1.4.1.post1
